Draft
move most of function to harmony
