+++ MoreOutsideInteraction.cs
@@ -71,68 +71,68 @@ namespace MoreOutsideInteraction
                 StreamReader sr = new StreamReader(fs);
                 string strLine = sr.ReadLine();
 
-                if (strLine == "False")
-                {
-                    deadFromOutside = false;
-                }
-                else
+                if (strLine == "True")
                 {
                     deadFromOutside = true;
                 }
+                else
+                {
+                    deadFromOutside = false;
+                }
 
                 strLine = sr.ReadLine();
 
-                if (strLine == "False")
-                {
-                    garbageFromOutside = false;
-                }
-                else
+                if (strLine == "True")
                 {
                     garbageFromOutside = true;
                 }
+                else
+                {
+                    garbageFromOutside = false;
+                }
 
                 strLine = sr.ReadLine();
 
-                if (strLine == "False")
-                {
-                    garbageToOutside = false;
-                }
-                else
+                if (strLine == "True")
                 {
                     garbageToOutside = true;
                 }
+                else
+                {
+                    garbageToOutside = false;
+                }
 
                 strLine = sr.ReadLine();
 
-                if (strLine == "False")
-                {
-                    crimeToOutside = false;
-                }
-                else
+                if (strLine == "True")
                 {
                     crimeToOutside = true;
                 }
+                else
+                {
+                    crimeToOutside = false;
+                }
 
                 strLine = sr.ReadLine();
 
-                if (strLine == "False")
-                {
-                    sickToOutside = false;
-                }
-                else
+                if (strLine == "True")
                 {
                     sickToOutside = true;
                 }
+                else
+                {
+                    sickToOutside = false;
+                }
 
                 strLine = sr.ReadLine();
 
-                if (strLine == "False")
+                if (strLine == "True")
                 {
-                    fireToOutside = false;
+                    fireToOutside = true;
                 }
                 else
                 {
-                    fireToOutside = true;
+                    fireToOutside = false;
                 }
 
                 sr.Close();
