+++ MoreOutsideInteraction.csproj
@@ -57,14 +57,18 @@
     </Reference>
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="CustomAI\CustomAmbulanceAI.cs" />
-    <Compile Include="CustomAI\CustomFireTruckAI.cs" />
-    <Compile Include="CustomAI\CustomGarbageTruckAI.cs" />
-    <Compile Include="CustomAI\CustomHearseAI.cs" />
-    <Compile Include="CustomAI\CustomHumanAI.cs" />
     <Compile Include="CustomAI\CustomOutsideConnectionAI.cs" />
     <Compile Include="CustomAI\CustomPlayerBuildingAI.cs" />
-    <Compile Include="CustomAI\CustomPoliceCarAI.cs" />
+    <Compile Include="Patch\HumanAIGetBuildingTargetPositionPatch.cs" />
+    <Compile Include="Patch\PoliceCarAIArriveAtTargetPatch.cs" />
+    <Compile Include="Patch\HearseAIArriveAtTargetPatch.cs" />
+    <Compile Include="Patch\GarbageTruckAIArriveAtTargetPatch.cs" />
+    <Compile Include="Patch\FireTruckAIArriveAtTargetPatch.cs" />
+    <Compile Include="Patch\PlayerBuildingAISimulationStepPatch.cs" />
+    <Compile Include="Patch\OutsideConnectionAISimulationStepPatch.cs" />
+    <Compile Include="Patch\FireTruckAIStartTransferPatch.cs" />
+    <Compile Include="Patch\AmbulanceAIStartTransferPatch.cs" />
+    <Compile Include="Patch\AmbulanceAIArriveAtTargetPatch.cs" />
     <Compile Include="Util\DebugLog.cs" />
     <Compile Include="Util\HarmonyDetours.cs" />
     <Compile Include="Loader.cs" />
