+++ Properties/AssemblyInfo.cs
@@ -32,5 +32,5 @@ using System.Runtime.InteropServices;
 //可以指定所有这些值，也可以使用“生成号”和“修订号”的默认值，
 // 方法是按如下所示使用“*”: :
 // [assembly: AssemblyVersion("1.0.*")]
-[assembly: AssemblyVersion("3.0.02.12")]
-[assembly: AssemblyFileVersion("3.0.02.12")]
+[assembly: AssemblyVersion("3.0.02.13")]
+[assembly: AssemblyFileVersion("3.0.02.13")]
