+++ Util/HarmonyDetours.cs
@@ -1,77 +1,22 @@
 ﻿using Harmony;
-using System.Reflection;
-using System;
-using UnityEngine;
-using MoreOutsideInteraction.CustomAI;
 
 namespace MoreOutsideInteraction.Util
 {
-    public static class HarmonyDetours
+    internal static class HarmonyDetours
     {
-        private static HarmonyInstance harmony = null;
-        private static void ConditionalPatch(this HarmonyInstance harmony, MethodBase method, HarmonyMethod prefix, HarmonyMethod postfix)
-        {
-            var fullMethodName = string.Format("{0}.{1}", method.ReflectedType?.Name ?? "(null)", method.Name);
-            if (harmony.GetPatchInfo(method)?.Owners?.Contains(harmony.Id) == true)
-            {
-                DebugLog.LogToFileOnly("Harmony patches already present for {0}" + fullMethodName.ToString());
-            }
-            else
-            {
-                DebugLog.LogToFileOnly("Patching {0}..." + fullMethodName.ToString());
-                harmony.Patch(method, prefix, postfix);
-            }
-        }
-
-        private static void ConditionalUnPatch(this HarmonyInstance harmony, MethodBase method, HarmonyMethod prefix = null, HarmonyMethod postfix = null)
-        {
-            var fullMethodName = string.Format("{0}.{1}", method.ReflectedType?.Name ?? "(null)", method.Name);
-            if (prefix != null)
-            {
-                DebugLog.LogToFileOnly("UnPatching Prefix{0}..." + fullMethodName.ToString());
-                harmony.Unpatch(method, HarmonyPatchType.Prefix);
-            }
-            if (postfix != null)
-            {
-                DebugLog.LogToFileOnly("UnPatching Postfix{0}..." + fullMethodName.ToString());
-                harmony.Unpatch(method, HarmonyPatchType.Postfix);
-            }
-        }
-
+        public const string ID = "pcfantasy.moreoutsideinteraction";
         public static void Apply()
         {
-            harmony = HarmonyInstance.Create("MoreOutsideInteraction");
-            //1
-            var outsideConnectionAISimulationStep = typeof(OutsideConnectionAI).GetMethod("SimulationStep", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Building).MakeByRefType()}, null);
-            var outsideConnectionAISimulationStepPostFix = typeof(CustomOutsideConnectionAI).GetMethod("OutsideConnectionAISimulationStepPostFix");
-            harmony.ConditionalPatch(outsideConnectionAISimulationStep,
-                null,
-                new HarmonyMethod(outsideConnectionAISimulationStepPostFix));
-            //2
-            var playerBuildingAISimulationStep = typeof(PlayerBuildingAI).GetMethod("SimulationStep", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Building).MakeByRefType(), typeof(Building.Frame).MakeByRefType() }, null);
-            var playerBuildingAISimulationStepPostFix = typeof(CustomPlayerBuildingAI).GetMethod("PlayerBuildingAISimulationStepPostFix");
-            harmony.ConditionalPatch(playerBuildingAISimulationStep,
-                null,
-                new HarmonyMethod(playerBuildingAISimulationStepPostFix));
-            Loader.HarmonyDetourInited = true;
+            var harmony = HarmonyInstance.Create(ID);
+            harmony.PatchAll();
+            Loader.HarmonyDetourFailed = false;
             DebugLog.LogToFileOnly("Harmony patches applied");
         }
 
         public static void DeApply()
         {
-            //1
-            var outsideConnectionAISimulationStep = typeof(OutsideConnectionAI).GetMethod("SimulationStep", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Building).MakeByRefType() }, null);
-            var outsideConnectionAISimulationStepPostFix = typeof(CustomOutsideConnectionAI).GetMethod("OutsideConnectionAISimulationStepPostFix");
-            harmony.ConditionalUnPatch(outsideConnectionAISimulationStep,
-                null,
-                new HarmonyMethod(outsideConnectionAISimulationStepPostFix));
-            //2
-            var playerBuildingAISimulationStep = typeof(PlayerBuildingAI).GetMethod("SimulationStep", BindingFlags.Public | BindingFlags.Instance, null, new Type[] { typeof(ushort), typeof(Building).MakeByRefType(), typeof(Building.Frame).MakeByRefType() }, null);
-            var playerBuildingAISimulationStepPostFix = typeof(CustomPlayerBuildingAI).GetMethod("PlayerBuildingAISimulationStepPostFix");
-            harmony.ConditionalUnPatch(playerBuildingAISimulationStep,
-                null,
-                new HarmonyMethod(playerBuildingAISimulationStepPostFix));
-            Loader.HarmonyDetourInited = false;
+            var harmony = HarmonyInstance.Create(ID);
+            harmony.UnpatchAll(ID);
             DebugLog.LogToFileOnly("Harmony patches DeApplied");
         }
     }
